Update mapblock.h
More work tmr

+++ src/mapblock.h
@@ -612,7 +612,7 @@ inline bool objectpos_over_limit(v3f p)
 
 inline bool blockpos_over_max_limit(v3s16 p)
 {
-	const s16 max_limit_bp = MAX_MAP_GENERATION_LIMIT / MAP_BLOCKSIZE;
+	const long max_limit_bp = MAX_MAP_GENERATION_LIMIT / MAP_BLOCKSIZE;
 	return p.X < -max_limit_bp ||
 		p.X >  max_limit_bp ||
 		p.Y < -max_limit_bp ||
